python-code/test_api/requirement.txt
pandas==2.3.0
python-dotenv==1.0.1
pinecone==5.3.1